+++ pkg/apis/ml/v1alpha1/extension_conditions.go
@@ -24,4 +24,6 @@ import api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 
 const (
 	ExtensionDeploymentFoundCondition api.ConditionType = "DeploymentFound"
+
+	ExtensionMetadataServiceValidCondition api.ConditionType = "MetadataServiceValid"
 )