+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - (Feature) Improve K8S Mock for UT
 - (Feature) (ML) Introduce basic Conditions
 - (Improvement) Raise memory requests for init containers to 50mi
+- (Feature) (ML) Metadata Service Implementation
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks